// web/src/api/models/evaluation-detail.ts
/* tslint:disable */
/* eslint-disable */
/**
 * ATRAG API
 * ATRAG API Documentation
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { EvaluationDetailConfig } from './evaluation-detail-config';
// May contain unused imports in some cases
// @ts-ignore
import type { EvaluationItem } from './evaluation-item';
// May contain unused imports in some cases
// @ts-ignore
import type { EvaluationStatus } from './evaluation-status';

/**
 * 
 * @export
 * @interface EvaluationDetail
 */
export interface EvaluationDetail {
    /**
     * 
     * @type {string}
     * @memberof EvaluationDetail
     */
    'id'?: string;
    /**
     * 
     * @type {string}
     * @memberof EvaluationDetail
     */
    'name'?: string;
    /**
     * 
     * @type {string}
     * @memberof EvaluationDetail
     */
    'collection_name'?: string;
    /**
     * 
     * @type {string}
     * @memberof EvaluationDetail
     */
    'question_set_name'?: string;
    /**
     * 
     * @type {EvaluationStatus}
     * @memberof EvaluationDetail
     */
    'status'?: EvaluationStatus;
    /**
     * 
     * @type {number}
     * @memberof EvaluationDetail
     */
    'average_score'?: number;
    /**
     * 
     * @type {EvaluationDetailConfig}
     * @memberof EvaluationDetail
     */
    'config'?: EvaluationDetailConfig;
    /**
     * 
     * @type {Array<EvaluationItem>}
     * @memberof EvaluationDetail
     */
    'items'?: Array<EvaluationItem>;
    /**
     * 
     * @type {string}
     * @memberof EvaluationDetail
     */
    'gmt_created'?: string;
}