// web/src/api/models/evaluation-item.ts
/* tslint:disable */
/* eslint-disable */
/**
 * ATRAG API
 * ATRAG API Documentation
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { EvaluationItemStatus } from './evaluation-item-status';

/**
 * 
 * @export
 * @interface EvaluationItem
 */
export interface EvaluationItem {
    /**
     * 
     * @type {string}
     * @memberof EvaluationItem
     */
    'id'?: string;
    /**
     * 
     * @type {string}
     * @memberof EvaluationItem
     */
    'evaluation_id'?: string;
    /**
     * 
     * @type {string}
     * @memberof EvaluationItem
     */
    'question_id'?: string;
    /**
     * 
     * @type {EvaluationItemStatus}
     * @memberof EvaluationItem
     */
    'status'?: EvaluationItemStatus;
    /**
     * 
     * @type {string}
     * @memberof EvaluationItem
     */
    'question_text'?: string;
    /**
     * 
     * @type {string}
     * @memberof EvaluationItem
     */
    'ground_truth'?: string;
    /**
     * 
     * @type {string}
     * @memberof EvaluationItem
     */
    'rag_answer'?: string;
    /**
     * 
     * @type {object}
     * @memberof EvaluationItem
     */
    'rag_answer_details'?: object;
    /**
     * 
     * @type {number}
     * @memberof EvaluationItem
     */
    'llm_judge_score'?: number;
    /**
     * 
     * @type {string}
     * @memberof EvaluationItem
     */
    'llm_judge_reasoning'?: string;
    /**
     * 
     * @type {string}
     * @memberof EvaluationItem
     */
    'gmt_created'?: string;
    /**
     * 
     * @type {string}
     * @memberof EvaluationItem
     */
    'gmt_updated'?: string;
}